+++ packages/editor/src/RichContentEditor/RichContentEditor.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { EditorState, convertFromRaw } from '@wix/draft-js';
 import Editor from 'draft-js-plugins-editor';
-import { get, includes, merge } from 'lodash';
+import { get, includes, merge, debounce } from 'lodash';
 import Measure from 'react-measure';
 import { translate } from 'react-i18next';
 import createEditorToolbars from './Toolbars';
@@ -32,6 +32,7 @@ class RichContentEditor extends Component {
     this.state = {
       editorState: this.getInitialEditorState(),
       theme: props.theme || {},
+      editorBounds: {},
     };
     this.refId = Math.floor(Math.random() * 9999);
 
@@ -63,9 +64,12 @@ class RichContentEditor extends Component {
       config,
       isMobile,
       setEditorState: this.setEditorState,
+      getEditorBounds: this.getEditorBounds,
       languageDir: getLangDir(locale),
     };
   };
+
+  getEditorBounds = () => this.state.editorBounds;
 
   initPlugins() {
     const {
@@ -97,6 +101,7 @@ class RichContentEditor extends Component {
       relValue,
       getEditorState: this.getEditorState,
       setEditorState: this.setEditorState,
+      getEditorBounds: this.getEditorBounds,
     });
     this.initEditorToolbars(pluginButtons, pluginTextButtons);
     this.pluginKeyBindings = initPluginKeyBindings(pluginTextButtons);
@@ -131,6 +136,7 @@ class RichContentEditor extends Component {
       theme: theme || {},
       getEditorState: this.getEditorState,
       setEditorState: this.setEditorState,
+      getEditorBounds: this.getEditorBounds,
       t,
       refId: this.refId,
       getToolbarSettings: config.getToolbarSettings,
@@ -226,7 +232,7 @@ class RichContentEditor extends Component {
   setEditor = ref => (this.editor = get(ref, 'editor', ref));
 
   updateBounds = editorBounds => {
-    this.subscriberPubsubs.forEach(pubsub => pubsub.set('editorBounds', editorBounds));
+    this.setState({ editorBounds });
   };
 
   renderToolbars = () => {
@@ -361,6 +367,8 @@ class RichContentEditor extends Component {
     return <style id="dynamicStyles">{css}</style>;
   };
 
+  onResize = debounce(({ bounds }) => this.updateBounds(bounds), 100);
+
   render() {
     const { isMobile } = this.props;
     const { theme } = this.state;
@@ -370,7 +378,7 @@ class RichContentEditor extends Component {
     });
     return (
       <Context.Provider value={this.contextualData}>
-        <Measure bounds onResize={({ bounds }) => this.updateBounds(bounds)}>
+        <Measure bounds onResize={this.onResize}>
           {({ measureRef }) => (
             <div dir={this.contextualData.languageDir}>
               <div style={this.props.style} ref={measureRef} className={wrapperClassName}>