+++ packages/plugin-map/src/createMapPlugin.js
@@ -5,7 +5,7 @@ import createToolbar from './toolbar';
 
 const createMapPlugin = (config = {}) => {
   const type = MAP_TYPE;
-  const { helpers, theme, t, [type]: settings = {}, isMobile, ...rest } = config;
+  const { helpers, theme, t, [type]: settings = {}, getEditorBounds, isMobile, ...rest } = config;
 
   return createBasePlugin({
     component: MapViewer,
@@ -13,6 +13,7 @@ const createMapPlugin = (config = {}) => {
     settings,
     theme,
     toolbar: createToolbar({
+      getEditorBounds,
       settings,
       helpers,
       theme,
@@ -20,6 +21,7 @@ const createMapPlugin = (config = {}) => {
       isMobile,
     }),
     helpers,
+    getEditorBounds,
     t,
     isMobile,
     ...rest,